backend/hangry_api/requirements.txt
argcomplete==2.0.0
asgiref==3.5.2
boto3==1.24.26
botocore==1.27.26
certifi==2022.6.15
cfn-flip==1.3.0
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
djangorestframework==3.13.1
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
jmespath==1.0.1
kappa==0.6.0
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.0
troposphere==4.0.2
urllib3==1.26.10
Werkzeug==2.1.2
wsgi-request-logger==0.4.6
zappa==0.54.1
